+++ lib/src/caches/async_dynamic_cache.dart
@@ -4,7 +4,7 @@ part of dart_dyncache;
 /// natively supports [Future] values.
 class AsyncBasicDynamicCache<K, V> implements DynamicCache<K, V> {
   late BaseDynamicCache<K, V> _wrappedCache;
-  final _waitingOnMap = <dynamic, Future<void>>{};
+  final _waitingOnMap = <dynamic, List<Future<V>>>{};
   int _waitingOnCount = 0;
 
   AsyncBasicDynamicCache(final BaseDynamicCache<K, V> wrappedCache) {
@@ -30,7 +30,7 @@ class AsyncBasicDynamicCache<K, V> implements DynamicCache<K, V> {
 
   /// Waits for all future values passed into [setAsync] to complete.
   Future<void> awaitAllCurrentFutureValues() async {
-    await Future.wait(_waitingOnMap.values);
+    await Future.wait(_waitingOnMap.values.expand((element) => element));
   }
 
   /// The current count of futures passed into [setAsync] that have not completed.
@@ -66,6 +66,9 @@ class AsyncBasicDynamicCache<K, V> implements DynamicCache<K, V> {
   @override
   set misses(int newMisses) => _wrappedCache.misses = newMisses;
 
+  @override
+  bool containsKey(dynamic key) => _wrappedCache.containsKey(key);
+
   /// If [key] is known to be associated with a future value, this method waits
   /// until that future is completed.
   ///
@@ -73,11 +76,15 @@ class AsyncBasicDynamicCache<K, V> implements DynamicCache<K, V> {
   /// is `true`, or if [setAsync] was called with an auxiliary [key] who's associated
   /// [AuxiliaryKeyManager.valueNeeded] is `true` this method will miss even
   /// if the future is known.
+  ///
+  /// If [setAsync] is called after [getAsync], the new future will not be awaited.
   Future<V?> getAsync(dynamic key) async {
+    V? futureValue;
     if (_waitingOnMap.containsKey(key)) {
-      await _waitingOnMap[key];
+      final results = await Future.wait(_waitingOnMap[key]!);
+      futureValue ??= results.last;
     }
-    return get(key, throwOnUnsafe: false);
+    return get(key, throwOnUnsafe: false) ?? futureValue;
   }
 
   @override
@@ -86,8 +93,28 @@ class AsyncBasicDynamicCache<K, V> implements DynamicCache<K, V> {
   @override
   void set(dynamic key, V value) => _wrappedCache.set(key, value);
 
+  /// If a value exists in the cache for [key], returns that value.
+  ///
+  /// Removes associated futures with [key] as soon as then complete.
   @override
-  V? remove(dynamic key) => _wrappedCache.remove(key);
+  V? remove(dynamic key) {
+    _waitingOnMap[key]?.forEach((element) {
+      element.then((value) => _wrappedCache.remove(key));
+    });
+    return _wrappedCache.remove(key);
+  }
+
+  List<Future<V>> _futureSetForKey(dynamic key) {
+    return _waitingOnMap.putIfAbsent(key, () => []);
+  }
+
+  void _removeFutureForKey(dynamic key, Future<V> future) {
+    final keySet = _waitingOnMap[key];
+    keySet?.remove(future);
+    if (keySet?.isEmpty ?? false) {
+      _waitingOnMap.remove(key);
+    }
+  }
 
   /// Associates [key] with the future result of [value].
   ///
@@ -102,42 +129,42 @@ class AsyncBasicDynamicCache<K, V> implements DynamicCache<K, V> {
   /// Any auxiliary key associated with [key] will miss with [getAsync] if
   /// the associated [AuxiliaryKeyManager.valueNeeded] is `true`.
   void setAsync(dynamic key, Future<V> value) {
-    final completer = Completer<void>();
+    final completer = Completer<V>();
     if (keys.contains(key)) {
-      _waitingOnMap[key] = completer.future;
+      _futureSetForKey(key).add(completer.future);
       final auxKeys =
           _wrappedCache.getAuxiliaryKeysForMainKeyAndValue(key, null);
       for (final auxKey in auxKeys) {
-        _waitingOnMap[auxKey] = completer.future;
+        _futureSetForKey(auxKey).add(completer.future);
       }
       _waitingOnCount++;
       value.then((value) {
         set(key, value);
-        _waitingOnMap.remove(key);
+        _removeFutureForKey(key, completer.future);
         for (final auxKey in auxKeys) {
-          _waitingOnMap.remove(auxKey);
+          _removeFutureForKey(auxKey, completer.future);
         }
-        completer.complete();
+        completer.complete(value);
         _waitingOnCount--;
       });
       return;
     }
     final mainKey = _wrappedCache.getMainKeyFromAuxiliaryKey(key, true);
     if (mainKey != null) {
-      _waitingOnMap[mainKey] = completer.future;
+      _futureSetForKey(mainKey).add(completer.future);
       final auxKeys =
           _wrappedCache.getAuxiliaryKeysForMainKeyAndValue(mainKey, null);
       for (final auxKey in auxKeys) {
-        _waitingOnMap[auxKey] = completer.future;
+        _futureSetForKey(auxKey).add(completer.future);
       }
       _waitingOnCount++;
       value.then((value) {
         set(mainKey, value);
-        _waitingOnMap.remove(mainKey);
+        _removeFutureForKey(mainKey, completer.future);
         for (final auxKey in auxKeys) {
-          _waitingOnMap.remove(auxKey);
+          _removeFutureForKey(auxKey, completer.future);
         }
-        completer.complete();
+        completer.complete(value);
         _waitingOnCount--;
       });
       return;
@@ -145,24 +172,24 @@ class AsyncBasicDynamicCache<K, V> implements DynamicCache<K, V> {
 
     // New insertion
     if (key is K) {
-      _waitingOnMap[key] = completer.future;
+      _futureSetForKey(key).add(completer.future);
       final auxKeys =
           _wrappedCache.getAuxiliaryKeysForMainKeyAndValue(key, null);
       for (final auxKey in auxKeys) {
-        _waitingOnMap[auxKey] = completer.future;
+        _futureSetForKey(auxKey).add(completer.future);
       }
       _waitingOnCount++;
       value.then((value) {
         set(key, value);
-        _waitingOnMap.remove(key);
+        _removeFutureForKey(key, completer.future);
         for (final auxKey in auxKeys) {
-          _waitingOnMap.remove(auxKey);
+          _removeFutureForKey(auxKey, completer.future);
         }
-        completer.complete();
+        completer.complete(value);
         _waitingOnCount--;
       });
     } else {
-      _waitingOnMap[key] = completer.future;
+      _futureSetForKey(key).add(completer.future);
       final otherKeys = <dynamic>{};
       final maybeMainKey =
           _wrappedCache.generateNewMainKeyFromAuxiliaryKey(key);
@@ -171,16 +198,19 @@ class AsyncBasicDynamicCache<K, V> implements DynamicCache<K, V> {
         otherKeys.addAll(_wrappedCache.getAuxiliaryKeysForMainKeyAndValue(
             maybeMainKey, null));
       }
+      for (final key in otherKeys) {
+        _futureSetForKey(key).add(completer.future);
+      }
       _waitingOnCount++;
       value.then((value) {
         final mainKey = _wrappedCache
             .generateNewMainKeyFromAuxiliaryKeyAndValue(key, value);
         set(mainKey, value);
-        _waitingOnMap.remove(key);
+        _removeFutureForKey(key, completer.future);
         for (final key in otherKeys) {
-          _waitingOnMap.remove(key);
+          _removeFutureForKey(key, completer.future);
         }
-        completer.complete();
+        completer.complete(value);
         _waitingOnCount--;
       });
     }