+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.2
+
+- Added `containsKey` method to caches.
+- `remove` method for `AsyncBasicDynamicCache` removes futures as well.
+- Downgraded the required verison
+
 ## 1.0.1
 
 - Added `clear` method to caches.